# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VETRA Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Add a document record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Contracts ───────────────────────────────────────────────────────────────
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "201":
          description: Created contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractUpdate"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"

  # ─── Daily Reports ────────────────────────────────────────────────────────────
  /daily-reports:
    get:
      operationId: listDailyReports
      tags: [daily-reports]
      summary: List daily reports
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of daily reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyReport"
    post:
      operationId: createDailyReport
      tags: [daily-reports]
      summary: Create a daily report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyReportInput"
      responses:
        "201":
          description: Created daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /daily-reports/{id}:
    get:
      operationId: getDailyReport
      tags: [daily-reports]
      summary: Get daily report by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"

  # ─── Meetings ────────────────────────────────────────────────────────────────
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List equipment
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Add equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"

  # ─── Inventory ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Document:
      type: object
      required:
        - id
        - name
        - type
        - size
        - projectId
        - projectName
        - uploadedBy
        - createdAt
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        size: { type: integer }
        projectId: { type: integer }
        projectName: { type: string }
        uploadedBy: { type: string }
        url: { type: ["string", "null"] }
        createdAt: { type: string }
    DocumentInput:
      type: object
      required:
        - name
        - type
        - size
        - projectId
        - uploadedBy
      properties:
        name: { type: string }
        type: { type: string }
        size: { type: integer }
        projectId: { type: integer }
        uploadedBy: { type: string }
        url: { type: string }

    # ─── Contract Schemas ─────────────────────────────────────────────────────
    Contract:
      type: object
      required:
        - id
        - name
        - contractor
        - value
        - status
        - projectId
        - projectName
        - startDate
        - endDate
        - createdAt
      properties:
        id: { type: integer }
        name: { type: string }
        contractor: { type: string }
        value: { type: number }
        status: { type: string }
        type: { type: ["string", "null"] }
        projectId: { type: integer }
        projectName: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        signedDate: { type: ["string", "null"] }
        createdAt: { type: string }
    ContractInput:
      type: object
      required:
        - name
        - contractor
        - value
        - projectId
        - startDate
        - endDate
      properties:
        name: { type: string }
        contractor: { type: string }
        value: { type: number }
        status: { type: string }
        type: { type: string }
        projectId: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
        signedDate: { type: string }
    ContractUpdate:
      type: object
      properties:
        name: { type: string }
        contractor: { type: string }
        value: { type: number }
        status: { type: string }
        type: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        signedDate: { type: string }

    # ─── Daily Report Schemas ─────────────────────────────────────────────────
    DailyReport:
      type: object
      required:
        - id
        - date
        - weather
        - progress
        - projectId
        - projectName
        - createdBy
        - createdAt
      properties:
        id: { type: integer }
        date: { type: string }
        weather: { type: string }
        temperature: { type: ["number", "null"] }
        progress: { type: number }
        workersOnSite: { type: integer }
        issues: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        projectId: { type: integer }
        projectName: { type: string }
        createdBy: { type: string }
        createdAt: { type: string }
    DailyReportInput:
      type: object
      required:
        - date
        - weather
        - progress
        - projectId
        - createdBy
      properties:
        date: { type: string }
        weather: { type: string }
        temperature: { type: number }
        progress: { type: number }
        workersOnSite: { type: integer }
        issues: { type: string }
        notes: { type: string }
        projectId: { type: integer }
        createdBy: { type: string }

    # ─── Meeting Schemas ──────────────────────────────────────────────────────
...
    Equipment:
      type: object
      required:
        - id
        - name
        - type
        - status
        - createdAt
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        model: { type: ["string", "null"] }
        serialNumber: { type: ["string", "null"] }
        status: { type: string }
        location: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        projectName: { type: ["string", "null"] }
        nextMaintenance: { type: ["string", "null"] }
        createdAt: { type: string }
    EquipmentInput:
      type: object
      required:
        - name
        - type
        - status
      properties:
        name: { type: string }
        type: { type: string }
        model: { type: string }
        serialNumber: { type: string }
        status: { type: string }
        location: { type: string }
        projectId: { type: integer }
        nextMaintenance: { type: string }
    EquipmentUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        model: { type: string }
        status: { type: string }
        location: { type: string }
        projectId: { type: integer }
        nextMaintenance: { type: string }

    # ─── Inventory Schemas ────────────────────────────────────────────────────
...