# lib/api-zod/src/generated/types/dailyReport.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * VETRA Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyReport {
  id: number;
  date: string;
  weather: string;
  /** @nullable */
  temperature?: number | null;
  progress: number;
  workersOnSite?: number;
  /** @nullable */
  issues?: string | null;
  /** @nullable */
  notes?: string | null;
  projectId: number;
  projectName: string;
  createdBy: string;
  createdAt: string;
}